requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mistralai==0.4.0
google-generativeai==0.3.2
pydantic>=2.5.2,<3.0.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
jsonschema==4.19.2
httpx==0.25.2
